Update Cody-Draft.md
I think that this essay is a really good start. There is a lot of information you can work with but as of right now it is brief, and seems to jump around quite a bit-which is understandable since Kelley is the only source used at the moment. You discuss a lot of different people which is what Kelley did. So it may be helpful to just pick a few important players and discuss them and their contribution to historiography.  I think once you are able to add in other sources and eliminate some people you have discussed, the essay will improve a lot. One thing that I do think will be interesting is if you discuss more on the Barbarian tribes. If you go back to the posts about the chapter in 'faces' folder you will see questions on the barbarians came up a few times. So maybe discussing why they chose to examine these barbarian tribes and develop their own histories, as well as how they went about it would be interesting. Overall I do think this essay is off to a good start.