Update universal-histories.md
Overall, I liked the writing, but the transitions were a little rough for a narrative. I would suggest expanding on and emphasizing the secular view on universal history. You have a great deal of writing on the Christian influence on Universal History, but your writing makes the theme of universal history seem like it is completely dominated by Christian thought. I'd also suggest taking a look at each section independently and try to get the flow from one idea to the next more smooth. It seems to be a problem we're both having, since we are both trying to cover vast time periods in short sections. I liked it though, your section on the christians was great.